+++ blueprints/flexberry-edit-form/files/__root__/controllers/__name__.js
@@ -2,6 +2,6 @@ import EditFormController from 'ember-flexberry/controllers/edit-form';
 
 export default EditFormController.extend({
   // Caption of this particular edit form.
-  title: '<%=caption%>',
-  parentRoute: '<%=parentRoute%>',
+  title: '<%=caption%>',<%if(parentRoute){%>
+  parentRoute: '<%=parentRoute%>',<%}%>
 });