+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Ember Flexberry Changelog
 ### New Features & improvements
 
+## 2016-06-23
+### Added
+FlexberryObjectlistviewComponent
+* Add support `modelProjection` by name
+* Add support promise for content
+
 ## 2016-06-22
 ### Added
 FlexberryCheckboxComponent:
 * Add class property for wrapper component.
+
+## 2016-06-21
+### Breaking changes
+FlexberryObjectlistview:
+* Change of mechanism for adding custom user buttons.
 
 ## 2016-06-14
 ### Fixed